Move all flat geometry functionality into flat.go
This is only to reduce the number of files at the top level, which makes
it easier for Github visitors to see README.md.
