--- flat_geom1.go
@@ -1,42 +1,0 @@
-package geom
-
-type geom1 struct {
-	geom0
-}
-
-// Coord returns the ith coord of g.
-func (g *geom1) Coord(i int) Coord {
-	return g.flatCoords[i*g.stride : (i+1)*g.stride]
-}
-
-// Coords unpacks and returns all of g's coordinates.
-func (g *geom1) Coords() []Coord {
-	return inflate1(g.flatCoords, 0, len(g.flatCoords), g.stride)
-}
-
-// NumCoords returns the number of coordinates in g.
-func (g *geom1) NumCoords() int {
-	return len(g.flatCoords) / g.stride
-}
-
-func (g *geom1) setCoords(coords1 []Coord) error {
-	var err error
-	g.flatCoords, err = deflate1(nil, coords1, g.stride)
-	return err
-}
-
-func (g *geom1) verify() error {
-	if g.stride != g.layout.Stride() {
-		return errStrideLayoutMismatch
-	}
-	if g.stride == 0 {
-		if len(g.flatCoords) != 0 {
-			return errNonEmptyFlatCoords
-		}
-	} else {
-		if len(g.flatCoords)%g.stride != 0 {
-			return errLengthStrideMismatch
-		}
-	}
-	return nil
-}
